requirements.txt
aiofiles==0.8.0
aiohappyeyeballs==2.4.3
aiohttp==3.9.0
aiosignal==1.3.1
anyio==4.6.1
asgiref==3.8.1
attrs==24.2.0
certifi==2024.8.30
charset-normalizer==2.1.1
click==8.1.7
fastapi==0.78.0
frozenlist==1.4.1
h11==0.14.0
idna==3.10
multidict==6.1.0
pip-system-certs==4.0
propcache==0.2.0
pydantic==1.10.18
python-dotenv==0.20.0
requests==2.28.1
sniffio==1.3.1
starlette==0.19.1
typing_extensions==4.12.2
urllib3==1.26.20
uvicorn==0.17.6
wrapt==1.16.0
yarl==1.15.2
